fix: oauth2 인증에 필요한 빈 CoreAppConfig 에서 api AppConfig로 이동
